refactor few main points (#11)
* refactor few main points
- authentication token now has 3 additional fields - companyId, institutionId, programId
- error handling is now more generic
- volunteer mapping now should handle cases when program manager or coordinator try to delete mapping
- updated tests to access the api instead of the use case function directly

* changed auth logic for deleteVolunteerFromRequest

* changed userDecoded to use `?` syntax

* added more tests and fixes to the model

* fixes to institution model

+++ src/app/volunteerRequest.ts
@@ -1,5 +1,8 @@
-import { VolunteerRequest } from '../models';
+import { AuthorizationError, CannotPerformOperationError, NotFoundError } from '../exc';
+import logger from '../logger';
+import { User, UserType, VolunteerRequest } from '../models';
 import { volunteerRequestRepository } from '../repos/volunteerRequestRepo';
+import { userDecoded } from './user';
 
 export const getRelevantAndOpenVolunteerRequests = async (): Promise<VolunteerRequest[]> => {
   return volunteerRequestRepository.relevantAndOpen();
@@ -13,6 +16,31 @@ export const getVolunteeRequestsByUser = async (userId: string): Promise<Volunte
   return volunteerRequestRepository.volunteerRequestsByVolunteerId(userId);
 };
 
-export const deleteVolunteerFromRequest = async (volunteerId: string, volunteerRequestId: number): Promise<void> => {
+const sameProgram = (userA: Partial<User>, userB: Partial<User>): boolean => {
+  return userA.programId === userB.programId;
+};
+
+export const deleteVolunteerFromRequest = async (
+  caller: userDecoded,
+  volunteerId: string,
+  volunteerRequestId: number
+): Promise<void> => {
+  if (caller.userType === UserType.VOLUNTEER && caller.userId !== volunteerId) {
+    throw new AuthorizationError('As a volunteer, you are not allowed to delete this volunteer');
+  }
+  const targetVolunteerRequest = await volunteerRequestRepository.findOneWithCreator(volunteerRequestId);
+  if (!targetVolunteerRequest) {
+    throw new NotFoundError('Volunteer request not found');
+  }
+  const requestCreator = targetVolunteerRequest.creator;
+  if (caller.userType === UserType.PROGRAM_MANAGER && !sameProgram(caller, requestCreator)) {
+    throw new AuthorizationError('As a program manager, you are not allowed to delete this volunteer');
+  }
+  if (caller.userType === UserType.PROGRAM_COORDINATOR && caller.userId !== requestCreator.id) {
+    throw new AuthorizationError('As a program coordinator, you are not allowed to delete this volunteer');
+  }
+  if (targetVolunteerRequest.startDate < new Date()) {
+    throw new CannotPerformOperationError('Cannot delete mapped volunteers from old request');
+  }
   await volunteerRequestRepository.deleteVolunteerFromRequest(volunteerRequestId, volunteerId);
 };
